pyproject.toml
[tool.poetry]
name = "dividas-api"
version = "0.1.0"
description = ""
authors = ["Gabriel <pgabrieldeveloper@outlook.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
sqlalchemy = {extras = ["asyncio"], version = "^1.4.42"}
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
python-dotenv = "^0.21.0"
greenlet = "^2.0.0.post0"
uvicorn = "^0.20.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^1.10.2"}
bcrypt = "^4.0.1"
python-multipart = "^0.0.5"
pytz = "^2022.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
